refactor: extract LLM parsing and dispatch helpers from analysis_tasks.ts
- Extract 13 functions and 2 types (170 lines) to lib/analysis_task_helpers.ts
- Includes: isObject, toNumber, toStringArray, parseBreakdown, unwrapLlmResult,
  parseLlmResult, extractKeywords, normalizeKeywords, stableHash,
  buildKeywordAnalysisId, buildAnalysisDispatchJobKey,
  buildAnalysisDispatchIdempotencyKey
- Add backward-compatible re-exports in analysis_tasks.ts
- analysis_tasks.ts reduced from 905 to 735 lines (19% reduction)
